+++ src/test/symbols/replacements.test.js
@@ -5,8 +5,6 @@ describe('Replace following symbols\n', () => {
 	let testCase = {
 		"(R) (R)": "® ®",
 		"(r) (r)": "® ®",
-		"(tm) (tm)": "™ ™",
-		"(TM) (TM)": "™ ™",
 		"+- +-": "± ±",
 		"-+ -+": "± ±",
 	};