+++ src/test/symbols/trademark.test.js
@@ -1,0 +1,19 @@
+import {fixTrademark} from "../../lib/symbols/trademark";
+import {fixTypos} from "../../typopo";
+import assert from 'assert';
+import Locale from "../../locale/locale";
+
+describe('Fix trademark ™\n', () => {
+	let testCase = {
+		"(tm)": "™",
+		"(TM)": "™",
+		"( tm )": "™",
+		"Company (tm)": "Company™",
+	};
+
+	Object.keys(testCase).forEach((key) => {
+		it("", () => {
+			assert.equal(fixTrademark(key, new Locale("en-us")), testCase[key]);
+		});
+	});
+});