+++ src/test/typopo.test.js
@@ -85,6 +85,8 @@ describe('Tests that all modules are plugged in', () => {
 		"Company (c)2017": "Company © 2017",
 		// exponents
 		"100 km3" : "100 km³",
+		// plus-minus
+		"+-": "±",
 		// sound recording copyright
 		"Company (p)2017": "Company ℗ 2017",
 		//registered trademark