+++ src/lib/symbols/plus-minus.js
@@ -1,0 +1,10 @@
+import Locale from "../../locale/locale";
+
+
+export function fixPlusMinus(string, locale) {
+	let pattern = "(\\+\\-)|(\\-\\+)";
+	let re = new RegExp(pattern, "g");
+	let replacement =  locale.plusMinus;
+
+	return string.replace(re, replacement);
+}