+++ src/typopo.js
@@ -1,10 +1,10 @@
 /*!
- * Typopo 2.1.2
+ * Typopo 2.1.3
  *
  * Copyright 2015-17 Braňo Šandala
  * Released under the MIT license
  *
- * Date: 2017-03-26
+ * Date: 2017-04-07
  */
 
 import Locale from "./locale/locale";
@@ -17,11 +17,13 @@ import {fixHyphen} from "./lib/punctuation/hyphen";
 import {fixDash} from "./lib/punctuation/dash";
 import {fixDoubleQuotesAndPrimes} from "./lib/punctuation/double-quotes";
 import {fixSingleQuotesPrimesAndApostrophes} from "./lib/punctuation/single-quotes";
-import {fixSymbols} from "./lib/symbols/replacements";
 import {fixMultiplicationSign} from "./lib/symbols/multiplication-sign";
 import {fixSectionSign} from "./lib/symbols/section-sign";
 import {fixCopyright} from "./lib/symbols/copyright";
 import {fixSoundRecordingCopyright} from "./lib/symbols/sound-recording-copyright";
+import {fixPlusMinus} from "./lib/symbols/plus-minus";
+import {fixRegisteredTrademark} from "./lib/symbols/registered-trademark";
+import {fixTrademark} from "./lib/symbols/trademark";
 import {fixExponents} from "./lib/symbols/exponents";
 import {fixAbbreviations} from "./lib/words/abbreviations";
 import {fixCase} from "./lib/words/case";
@@ -62,15 +64,17 @@ export function fixTypos(string, locale, configuration) {
 	string = fixPeriod(string);
 	string = fixDash(string, currentLocale);
 	string = fixHyphen(string, currentLocale);
-	string = fixDoubleQuotesAndPrimes(string, currentLocale);
 	string = fixSingleQuotesPrimesAndApostrophes(string, currentLocale);
+	string = fixDoubleQuotesAndPrimes(string, currentLocale);
 
 	// symbols
-	string = fixSymbols(string);
 	string = fixMultiplicationSign(string, currentLocale);
 	string = fixSectionSign(string, currentLocale);
 	string = fixCopyright(string, currentLocale);
 	string = fixSoundRecordingCopyright(string, currentLocale);
+	string = fixPlusMinus(string, currentLocale);
+	string = fixRegisteredTrademark(string, currentLocale);
+	string = fixTrademark(string, currentLocale);
 	string = fixExponents(string, currentLocale);
 
 	// words